Use byte instead of ushort for voice sequence numbers
Saves 1 byte per voice packet on the wire. Byte (0-255) wraps every
~5s at 50 pps, which is fine since the jitter buffer only looks 32
slots ahead. Uses sbyte cast for correct wrapping arithmetic.

+++ Basis Server/BasisServerTests/SerializableMessageTests.cs
@@ -315,7 +315,7 @@ namespace BasisServerTests
             var result = new AudioSegmentDataMessage();
             result.Deserialize(reader);
 
-            Assert.Equal((ushort)42, result.SequenceNumber);
+            Assert.Equal((byte)42, result.SequenceNumber);
             Assert.Equal(5, result.TotalPlayedInSilence);
             Assert.Equal(new byte[] { 1, 2, 3, 4 }, result.buffer);
             Assert.Equal(4, result.LengthUsed);
@@ -339,7 +339,7 @@ namespace BasisServerTests
             var result = new AudioSegmentDataMessage();
             result.Deserialize(reader);
 
-            Assert.Equal((ushort)100, result.SequenceNumber);
+            Assert.Equal((byte)100, result.SequenceNumber);
             Assert.Equal(10, result.TotalPlayedInSilence);
             Assert.Equal(0, result.LengthUsed);
         }
