+++ Basis/Packages/com.basis.framework/Networking/Recievers/BasisSequencedVoiceData.cs
@@ -1,12 +1,12 @@
 [System.Serializable]
 public struct BasisSequencedVoiceData
 {
-    public ushort SequenceNumber;
+    public byte SequenceNumber;
     public byte[] Array;
     public int Length;
     public bool IsInsertedSilence;
 
-    public BasisSequencedVoiceData(ushort sequenceNumber, byte[] array, int length, bool isInsertedSilence)
+    public BasisSequencedVoiceData(byte sequenceNumber, byte[] array, int length, bool isInsertedSilence)
     {
         SequenceNumber = sequenceNumber;
         Array = array;
